initial commit
This is the initial commit for the zscaler security preview tests run in a headless chromium with puppeteer as a NodeJS Lambda function. http://securitypreview.zscaler.com can be run by users with web browsers, so this is simply a way to run the tests as a Lambda function. This also includes the default Zscaler root certificate for SSL inspection in ZIA.